etl_test_framework/requirements.txt
pytest==8.1.1
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-ordering==0.6
pandas==2.2.1
sqlalchemy==2.0.29
pymysql==1.1.0
pyyaml==6.0.1
python-dotenv==1.0.1
great-expectations==0.18.12
jinja2==3.1.3
openpyxl==3.1.2
colorlog==6.8.2
